feat: better progress display for upload script
Display the progress as percentage on a single line that refreshes.

In order to see the progress and have the results logged into a file
use `tee` and use Python's unbuffered option:

    python3 -u ./upload_video.py | tee /tmp/upload.log

Thanks @tkardi for the idea of the progress display.

Also print the successful upload as JSON, so that the results can
be processed easily (e.g. to check if all files were really uploaded).

The resulting file contains also the progress. To see the JSON data only
use `col`. E.g.:

    cat /tmp/upload.log | col | /tmp/data.log